backend/requirements.txt
fastapi
uvicorn
python-multipart
pymongo
python-dotenv
easyocr
numpy
pymupdf
opencv-python-headless
Pillow